[Chore] Refactor code
* Refactor code to allow modern framework support
* Update pipeline to take Wikipedia data as training set
