+++ run.sh
@@ -1,0 +1,16 @@
+#!/bin/bash
+
+#SBATCH --job-name="pr"
+#SBATCH --output="%x.o%j"
+#SBATCH --nodes=1
+#SBATCH --ntasks=1
+#SBATCH --gres=gpu
+#SBATCH --cpus-per-task=4
+#SBATCH --time=48:00:00
+#SBATCH --mem=0
+#SBATCH --mail-user=jm3743@georgetown.edu
+#SBATCH --mail-type=END,FAIL
+
+source env.sh
+python prep-pr-data.py
+python main.py
